requirements.txt
flask==3.0.3
selenium==4.21.0
webdriver-manager==4.0.1
python-dotenv==1.0.0